requirements.txt
flask 
openai 
requests 
Werkzeug 
pdfrw 
reportlab
PyMuPDF
frontend
python-dotenv

#!/bin/bash
# sudo apt-get install graphviz
